+++ frontend/packages/client/src/components/index.js
@@ -46,7 +46,8 @@ export { default as Form } from './common/Form';
 export { default as Input } from './Input';
 export { default as TooltipMessage } from './TooltipMessage';
 export { default as FadeInOut } from './FadeInOut';
-export { default as ErrorModal } from './ErrorModal';
+export { default as ErrorModal } from './Modals/ErrorModal';
+export { default as Modal } from './Modals/Modal';
 export { default as StatusPill } from './StatusPill';
 export { default as StyledStatusPill } from './StyledStatusPill';
 export { default as FilterPill } from './FilterPill';
